色の受け渡しロジック完成
