+++ src/app/api/game/event2a/route.ts
@@ -1,0 +1,3 @@
+// ここには本来はバックエンドからやってくるデータを記入
+// 今回は仮で作成中
+// event1/route.tsを参考に見てね
