+++ src/app/api/game/type.ts
